# backend/docs/swagger/openapi.yml
openapi: 3.0.3
info:
  title: ERP Distribution - API Gateway
  version: 1.0.0
  description: |
    Specification OpenAPI centralisee exposee via l'API Gateway.
    Les clients consomment principalement ces routes /api/v1/*.
servers:
  - url: http://localhost:4000
    description: API Gateway locale
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: erp_session
paths:
...
  /api/v1/catalog/products:
    get:
      summary: Lister les produits (via gateway)
      tags: [Catalog]
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: category
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Liste des produits
        '403':
          description: Permission products.read requise
    post:
      summary: Creer un produit (via gateway)
      tags: [Catalog]
      security:
        - bearerAuth: []
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                product_id:
                  type: string
                  description: Optionnel (auto-genere si absent, ex PRD-AU-00001000)
                product_name:
                  type: string
                supplier_id:
                  type: string
                supplier_name:
                  type: string
                  description: Alternative a supplier_id
              required: [product_name]
      responses:
        '201':
          description: Produit cree
        '403':
          description: Permission products.create requise
  /api/v1/catalog/products/{productId}:
    get:
      summary: Detail produit (via gateway)
      tags: [Catalog]
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Detail produit
        '404':
          description: Non trouve
        '403':
          description: Permission products.read requise
    patch:
      summary: Mettre a jour un produit (via gateway)
      tags: [Catalog]
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Produit mis a jour
        '403':
          description: Permission products.update requise
    delete:
      summary: Supprimer un produit (via gateway)
      tags: [Catalog]
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Produit supprime
        '403':
          description: Permission products.delete requise
  /api/v1/catalog/products/{productId}/stock:
    patch:
      summary: Mettre a jour le stock (via gateway)
      tags: [Catalog]
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Stock mis a jour
        '403':
          description: Permission products.update_stock requise
...